refactor(gui-app): extract shared setup-terminal tab register hook
The chat and terminal-agent tab-register drivers ran the identical
registration effect (running-gate -> sessionId guard -> registerOnce ->
openTileInBackgroundTab), differing only in how they source the worktree
binding. Factor that body into a shared
`useRegisterSetupTerminalTabsFromBinding` hook so the two can't drift; each
driver now just sources its binding (chat from its live store, the tui
agent from the polled `worktree.getBinding`) and delegates.

The `openTileInBackgroundTab` eslint exemption moves to the single shared
hook - the one site that actually calls it - instead of being duplicated
per driver.

Addresses CodeRabbit review feedback on traycerai/traycer#313.
